refactor(plugins): split plugin type contracts
