+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-aaab273f4a65e33165556b0a9b35a4de1bebbef4786214de59de9efb32ffb1a3  plugin-sdk-api-baseline.json
-05dc60f0b6a4189a11f6409da696d55c9ee246ae7e89b775f23f6dc4ac874e17  plugin-sdk-api-baseline.jsonl
+fcf33e39751fdeebf70e4151372171b57de1b7318aeafa9478bbe512fc4634b8  plugin-sdk-api-baseline.json
+86c50a31e445c4c801920c72922f7cd6b0ee9bee241c2fdc1ab6d61964a9aedc  plugin-sdk-api-baseline.jsonl
