+++ src/plugins/worker-provider.types.ts
@@ -1,0 +1,69 @@
+import type { SecretRef } from "../config/types.secrets.js";
+import type { PluginJsonValue } from "./host-hooks.js";
+
+export type WorkerProfile = Readonly<Record<string, PluginJsonValue>>;
+
+/** SSH endpoint material returned by a worker provider after provisioning. */
+export type WorkerSshEndpoint = {
+  host: string;
+  port: number;
+  user: string;
+  /** OpenSSH public host-key line obtained from trusted provisioning output. */
+  hostKey: string;
+  /** Secret reference only; providers must never return plaintext key material. */
+  keyRef: SecretRef;
+};
+
+/** Resolved SSH client identity. Providers may return a local path or ephemeral material. */
+export type WorkerSshIdentity =
+  | { kind: "path"; path: string }
+  | { kind: "material"; contents: string };
+
+/** Durable context supplied when a worker provider resolves the identity it minted. */
+export type WorkerSshIdentityRequest = {
+  leaseId: string;
+  profile: WorkerProfile;
+  keyRef: SecretRef;
+};
+
+/** Durable lease identity and endpoint returned by a successful provision operation. */
+export type WorkerLease = {
+  leaseId: string;
+  ssh: WorkerSshEndpoint;
+};
+
+/** Authoritative inspection result for an already-known worker lease. */
+export type WorkerLeaseStatus =
+  | { status: "active" }
+  | { status: "destroyed" }
+  | { status: "unknown" };
+
+/** Permanent provider rejection recorded as a terminal worker failure. */
+export class WorkerProviderError extends Error {
+  readonly code = "invalid_profile";
+
+  constructor(message: string) {
+    super(message);
+    this.name = "WorkerProviderError";
+  }
+}
+
+/** Cloud-worker lifecycle capability registered by a plugin. */
+export type WorkerProvider = {
+  id: string;
+  /**
+   * Provision or adopt the lease for this operation id.
+   * Repeating the same operation id must be idempotent across gateway restarts.
+   */
+  provision: (profile: WorkerProfile, operationId: string) => Promise<WorkerLease>;
+  /** Throws on transient/indeterminate failures; `unknown` means authoritative absence. */
+  inspect: (lease: { leaseId: string; profile: WorkerProfile }) => Promise<WorkerLeaseStatus>;
+  /**
+   * Resolves provider-owned dynamic identities. When absent, the gateway uses its generic
+   * SecretRef resolver; when present, failures are authoritative and never fall back.
+   */
+  resolveSshIdentity?: (request: WorkerSshIdentityRequest) => Promise<WorkerSshIdentity>;
+  renew?: (leaseId: string) => Promise<void>;
+  /** Idempotent; resolves only after the provider can prove teardown. */
+  destroy: (lease: { leaseId: string; profile: WorkerProfile }) => Promise<void>;
+};
